ext2: fix build

+++ kernel/fs/ext2.c
@@ -318,7 +318,7 @@ static ext2_dir_t* direntry_ext2(
 }
 
 static fs_error_t get_child_ext2(fs_node_t* node, char* name, fs_node_t** out);
-static fs_error_t read_ext2(fs_node_t* node, uint32_t offset, uint8_t* buffer, uint32_t size);
+static fs_error_t read_ext2(fs_node_t* node, size_t offset, uint8_t* buffer, uint32_t size);
 static fs_error_t list_ext2(fs_node_t* node, fs_list_entry_t** eout);
 
 static fs_error_t node_from_file(
